data-science/requirements.txt
pandas
scikit-learn
lightgbm
neo4j
psycopg2-binary
scikit-learn